Create dew_source_code_v1_1_2.py
Updated source code for DEW4SLHAv1.1.2. This includes checks for if the maximum RGE scale for presenting soft values exceeds 2*sqrt(mt1*mt2), where mt1,2 are the pole masses of the stop squarks, then radiative corrections will not be minimized and thus results may not be accurate.

This can be addressed in your spectrum generator with the SLHA input block MODSEL, flags number 11 and 12. Most generators have the default max scale value (entry 12) as the geometric mean of the stop masses. If a user alters the inputs to too large of a scale, unscrupulous results may follow if using DEW4SLHA.