Created sample points for each model
A sample point is included for the CMSSM, NUHM1, NUHM2, NUHM3, NUHM4, and the pMSSM30 for users to test. DO NOT CHANGE THE VALUES IN THESE FILES, AS THEY ARE CALCULATED QUANTITIES FROM SOFTSUSY. Generate your own file if you wish to test a different BM point in any of these models.